+++ src/asp/Tools/pc_align.cc
@@ -145,7 +145,7 @@ void handle_arguments( int argc, char *argv[], Options& opt ) {
     ("max-num-source-points",    po::value(&opt.max_num_source_points)->default_value(100000),
                                  "Maximum number of (randomly picked) source points to use (after discarding gross outliers).")
     ("alignment-method",         po::value(&opt.alignment_method)->default_value("point-to-plane"),
-                                 "The type of iterative closest point method to use. [point-to-plane, point-to-point, similarity-point-to-point, fgr, least-squares, similarity-least-squares]")
+                                 "The type of iterative closest point method to use. [point-to-plane, point-to-point, similarity-point-to-point, similarity-point-to-plane, fgr, least-squares, similarity-least-squares]")
     ("highest-accuracy",         po::bool_switch(&opt.highest_accuracy)->default_value(false)->implicit_value(true),
                                  "Compute with highest accuracy for point-to-plane (can be much slower).")
     ("csv-format",               po::value(&opt.csv_format_str)->default_value(""), asp::csv_opt_caption().c_str())
@@ -273,20 +273,22 @@ void handle_arguments( int argc, char *argv[], Options& opt ) {
   if (opt.alignment_method != "point-to-plane"            &&
       opt.alignment_method != "point-to-point"            &&
       opt.alignment_method != "similarity-point-to-point" &&
+      opt.alignment_method != "similarity-point-to-plane" &&
       opt.alignment_method != "fgr"                       &&
       opt.alignment_method != "least-squares"             &&
       opt.alignment_method != "similarity-least-squares"
       )
     vw_throw( ArgumentErr() << "Only the following alignment methods are supported: "
 	      << "point-to-plane, point-to-point, similarity-point-to-point, "
-	      << "fgr, least-squares, and similarity-least-squares.\n"
+              << "similarity-point-to-plane, fgr, least-squares, and similarity-least-squares.\n"
 	      << usage << general_options );
 
   if (opt.alignment_method != "point-to-plane"            &&
       opt.alignment_method != "point-to-point"            &&
       opt.alignment_method != "similarity-point-to-point" &&
+      opt.alignment_method != "similarity-point-to-plane" &&
       opt.compute_translation_only) {
-    vw_throw( ArgumentErr() << "The option --compute-translation-only is only applicable to point-to-plane, point-to-point, and similarity-point-to-point alignment.\n"
+    vw_throw( ArgumentErr() << "The option --compute-translation-only is only applicable to point-to-plane, point-to-point, similarity-point-to-point, and similarity-point-to-plane alignment.\n"
 	      << usage << general_options );
   }
   
@@ -1415,7 +1417,8 @@ int main( int argc, char *argv[] ) {
         T = fgr_alignment(source_point_cloud, ref_point_cloud, opt);
       } else if (opt.alignment_method == "point-to-plane" ||
                  opt.alignment_method == "point-to-point" ||
-                 opt.alignment_method == "similarity-point-to-point") {
+                 opt.alignment_method == "similarity-point-to-point" ||
+                 opt.alignment_method == "similarity-point-to-plane") {
         // Use libpointmatcher
         T = icp(source_point_cloud, ref_point_cloud, Id,
 		opt.compute_translation_only);